further additions to the models classes
